Addition of POS terminals (#297)
* Added Terminals class

For Mollie POS transactions, the terminals class needed to be added to al required classes and objects

* updated summary

* updated the documentation

Some of the classes were lacking proper documentation, this push corrects these. Also added a link to the documentation of the classes. Attention, documentation on Mollie's website needs to be updated.

* Update PaymentMethod.cs

Updated POS to PointOfSale in the PaymentMethod to keep the same syntax